HYD-6745 Move common test utilities to stand-alone package
In order to make common utility test routines usable across number of chroma
top-level directories and/or from chroma packages which may not have access to
the full chroma source directory structure, a new chroma package called
chroma-common-test should be created which contains library of common test
utilities.

This package is created from within the chroma-common.spec file and consists of
files within the chroma-common/chroma_common/test/ directory.

  - Move CommandCaptureTestCase to chroma_common/test
  - Create ImlUnitTestCase in chroma_common/test
  - Make AgentUnitTestCase inherit from ImlUnitTestCase
  - Use multiple inheritance for modules that require CommandCaptureTestCase
    and AgentUnitTestCase
  - Create sub-package chroma-common-test in chroma-common.spec
  - Make chroma-manager-integration-tests require chroma-common-test
  - Consolidate test_manage_target.py into test_manager_targets.py
  - Amend and update unit tests
  - Fix Copyright notices
