+++ chroma-agent/tests/actions_plugins/test_manage_client_mounts.py
@@ -2,7 +2,7 @@ import os
 from StringIO import StringIO
 from mock import patch, mock_open
 
-from tests.command_capture_testcase import CommandCaptureTestCase
+from chroma_common.test.command_capture_testcase import CommandCaptureTestCase
 
 
 class TestClientMountManagement(CommandCaptureTestCase):
