+++ chroma-agent/tests/actions_plugins/test_manage_copytool.py
@@ -2,11 +2,12 @@ import tempfile
 import shutil
 import mock
 
-from tests.command_capture_testcase import CommandCaptureTestCase, CommandCaptureCommand
+from chroma_common.test.command_capture_testcase import CommandCaptureTestCase, CommandCaptureCommand
 from chroma_agent.action_plugins.manage_copytool import start_monitored_copytool, stop_monitored_copytool, configure_copytool, unconfigure_copytool, update_copytool, list_copytools, _copytool_vars
+from tests.lib.agent_unit_testcase import AgentUnitTestCase
 
 
-class TestCopytoolManagement(CommandCaptureTestCase):
+class TestCopytoolManagement(CommandCaptureTestCase, AgentUnitTestCase):
     def setUp(self):
         super(TestCopytoolManagement, self).setUp()
