+++ chroma-agent/tests/common/blockdevices/test_blockdevice_linux.py
@@ -2,7 +2,7 @@ import mock
 
 from chroma_common.blockdevices.blockdevice_linux import BlockDeviceLinux
 from tests.common.blockdevices.blockdevice_base_tests import BaseTestBD
-from tests.command_capture_testcase import CommandCaptureCommand
+from chroma_common.test.command_capture_testcase import CommandCaptureCommand
 
 
 class TestBlockDeviceLinux(BaseTestBD.BaseTestBlockDevice):
