+++ chroma-agent/tests/device_plugins/linux/test_device_mapper.py
@@ -1,4 +1,4 @@
-from tests.command_capture_testcase import CommandCaptureTestCase, CommandCaptureCommand
+from chroma_common.test.command_capture_testcase import CommandCaptureTestCase, CommandCaptureCommand
 from chroma_agent.device_plugins.linux_components.device_mapper import DmsetupTable
