+++ chroma-agent/tests/device_plugins/linux/test_zfs.py
@@ -3,7 +3,7 @@ from mock import patch
 from chroma_agent.device_plugins.linux import ZfsDevices
 from chroma_agent.device_plugins.linux_components.block_devices import BlockDevices
 from tests.device_plugins.linux.test_linux import LinuxAgentTests
-from tests.command_capture_testcase import CommandCaptureTestCase, CommandCaptureCommand
+from chroma_common.test.command_capture_testcase import CommandCaptureTestCase, CommandCaptureCommand
 
 
 class TestZfs(LinuxAgentTests, CommandCaptureTestCase):
