+++ chroma-agent/tests/lib/agent_unit_testcase.py
@@ -1,14 +1,8 @@
-import mock
-
-from django.utils import unittest
-
 from chroma_common.lib.agent_rpc import agent_error, agent_result_ok
+from chroma_common.test.iml_unit_testcase import ImlUnitTestCase
 
 
-class AgentUnitTestCase(unittest.TestCase):
-    def setUp(self):
-        self.addCleanup(mock.patch.stopall)
-
+class AgentUnitTestCase(ImlUnitTestCase):
     def assertAgentOK(self, value):
         self.assertEqual(value, agent_result_ok)
