+++ chroma-agent/tests/test_conf_params.py
@@ -1,6 +1,6 @@
 
 from chroma_agent.action_plugins.manage_conf_params import set_conf_param
-from tests.command_capture_testcase import CommandCaptureTestCase
+from chroma_common.test.command_capture_testcase import CommandCaptureTestCase
 
 
 class TestConfParams(CommandCaptureTestCase):
