+++ chroma-agent/tests/test_fail_node.py
@@ -1,6 +1,6 @@
 
 from chroma_agent.action_plugins.manage_fail_node import fail_node
-from tests.command_capture_testcase import CommandCaptureTestCase, CommandCaptureCommand
+from chroma_common.test.command_capture_testcase import CommandCaptureTestCase, CommandCaptureCommand
 
 
 class TestConfParams(CommandCaptureTestCase):
