+++ chroma-agent/tests/test_shutdown_reboot.py
@@ -2,7 +2,7 @@ from mock import patch
 
 from chroma_agent.action_plugins.manage_node import shutdown_server, reboot_server
 from chroma_agent.device_plugins.action_runner import CallbackAfterResponse
-from tests.command_capture_testcase import CommandCaptureTestCase
+from chroma_common.test.command_capture_testcase import CommandCaptureTestCase
 
 
 class TestServerShutdownAndReboot(CommandCaptureTestCase):
