+++ chroma-common/chroma-common.spec
@@ -1,7 +1,7 @@
 %{!?name: %define name chroma-common}
-%{?!version: %define version %(%{__python2} -c "from chroma_common import version; sys.stdout.write(version())")}
+%{?!version: %define version %(%{__python} -c "from chroma_common import version; sys.stdout.write(version())")}
 %{?!package_release: %define package_release 1}
-%{?!python2_sitelib: %define python2_sitelib %(%{__python2} -c "from distutils.sysconfig import get_python_lib; import sys; sys.stdout.write(get_python_lib())")}
+%{?!python_sitelib: %define python_sitelib %(%{__python} -c "from distutils.sysconfig import get_python_lib; import sys; sys.stdout.write(get_python_lib())")}
 
 Summary: Chroma Common
 Name: %{name}
@@ -17,19 +17,36 @@ Vendor: Intel(R) Corporation
 %description
 Common library containing routines used by both agent and manager.
 
+%package test
+Summary: Common test utilities for Intel Manager for Lustre tests
+Group: Development
+Requires: %{name} = %{version}-%{release}
+%description test
+This package contains shared test utilities used in the test framework for Manager for Lustre.
+
 %prep
 %setup -n %{name}-%{version}
 
 %build
-%{__python2} setup.py build
+%{__python} setup.py build
 
 %install
 rm -rf %{buildroot}
-%{__python2} setup.py install --skip-build --install-lib=%{python2_sitelib} --root=%{buildroot}
+%{__python} setup.py install --skip-build --install-lib=%{python_sitelib} --root=%{buildroot}
 mkdir -p $RPM_BUILD_ROOT/usr/sbin/
+
+touch test.files
+for test_file in $(find -L $RPM_BUILD_ROOT%{python_sitelib}/chroma_common/test/ -type f -name '*.py*'); do
+  install_file=${test_file/$RPM_BUILD_ROOT\///}
+  echo "${install_file}" >> test.files
+done
 
 %clean
 rm -rf %{buildroot}
 
 %files
-%{python2_sitelib}/*
+%exclude %{python_sitelib}/chroma_common/test
+%{python_sitelib}/*
+
+%files -f test.files test
+%defattr(-,root,root)
